pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scrappy-nepse"
version = "0.1.0"
description = "Clean NEPSE market and IPO data scraper"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "python-dateutil>=2.9.0",
  "requests>=2.31.0",
  "nepse-scraper>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
  "black>=24.8.0",
  "pre-commit>=3.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "UP"]
ignore = ["E501", "N818"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.setuptools.packages.find]
include = ["scraper*"]
